40º - PASO A CENTIMOS
Se ha pasado toda la logica de precios a centimos para trabajar mejor y que no haya casos de error de cálculos.

Se han modificado el muestreo en las vistas para que se muestren en euros

la base de datos se ha modificado para que guarde ints

Controladores que hacían cálculos con el precio se han simplificado (por ejemplo el controlador de stripe)

+++ database/migrations/2025_01_02_231114_create_seller_carts_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
             $table->foreignId('user_id')->nullable()->constrained('users')->onDelete('cascade');
             $table->string('token')->unique()->nullable();
             $table->integer('quantity');
-            $table->decimal('total_price', 10, 2)->default(0);
+            $table->integer('total_price')->default(0);
             $table->timestamps();
         });
     }
