+++ database/migrations/2025_01_26_112708_create_orders_table.php
@@ -10,7 +10,7 @@ return new class extends Migration {
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table->string('status');
-            $table->decimal('total_price');
+            $table->integer('total_price');
             $table->foreignId('buyer_id')->constrained('users');
             $table->foreignId('seller_id')->constrained('users');
             $table->string('shipment_number')->nullable();
